Change destroy timeout from 3 minutes to 5 minutes.
While destroying a cluster through Concourse, my destroy job failed
due to timeout... but the cluster was gone when I checked the Google
Cloud Platform UI. Looking at the VM history, it took exactly 3
minutes and 18.051 seconds to finish the destroy. Let's up our timeout
to 5 minutes so we get less false positives. I was thinking increasing
it to 10 minutes but I'll let future users decide if that's necessary
or not.